% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{corr_mat}
\alias{corr_mat}
\title{An nbapackage function}
\usage{
corr_mat(year)
}
\arguments{
\item{year}{The year from which you want the correlation matrix}
}
\description{
This function will output the correlation matrix of all continuous variables from a given year
}
\examples{
corr_mat()
}
\keyword{correlation}
\keyword{matrix}
