% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{highest_points}
\alias{highest_points}
\title{An nbapackage function}
\usage{
highest_points(year)
}
\arguments{
\item{year}{The year from which you want the top scorer.}
}
\description{
This function will give out the player who scored the most amount of points in the NBA for the selected year.
}
\examples{
highest_points()
}
\keyword{scorer}
\keyword{top}
